TOX: Add ox environment
We add tox environment and split frames into frames and storage.
