+++ phmu/__init__.py
@@ -1,4 +1,4 @@
 #!/bin/env python3
 # -*- coding: utf-8 -*- 
 
-__all__ = ['frames']
+__all__ = ['frames', 'storage']
