+++ setup.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+
+from distutils.core import setup
+
+setup(name='phmu',
+      version='0.0.1',
+      description='Photometry Utilities',
+      author='author',
+      author_email='author@example.com',
+      url='https://github.com/jara101/phmu/',
+      packages=['phmu'],
+    )
