+++ tox.ini
@@ -1,0 +1,13 @@
+[tox]
+envlist = py34
+
+[testenv]
+install_command = pip install {opts} {packages}
+deps = -r{toxinidir}/requirements.txt
+
+[testenv:pep8]
+commands = flake8
+
+[flake8]
+show-source = True
+exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
